A gaming website.
My first gaming website , and also my first website where I used card. I created this website in 3 days. It was a good experience for me, but my bad luck is that this website is not responsive. But it's ok. The most important thing for me that I code this website fully with my android device.
Thanks for reading.